Added keyboard navigation to presentation mode Cleaned up existing logic to separate files Removed redundant or incorrect code
