+++ package.json
@@ -89,7 +89,8 @@
     "vite": "^7.1.5",
     "vite-plugin-pwa": "^1.0.3",
     "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^3.2.4"
+    "vitest": "^3.2.4",
+    "workbox-window": "^7.3.0"
   },
   "jest": {
     "clearMocks": false
