+++ src/components/Songs/SongDisplay/SongDisplay.scss
@@ -185,7 +185,7 @@
 }
 
 .navigation-button {
-  flex: 1;
+  flex: 1 1 45%;
   max-width: 500px;
   min-height: 42px;
 }
