+++ tsconfig.json
@@ -1,6 +1,12 @@
 {
   "compilerOptions": {
-    "types": ["node", "@testing-library/jest-dom", "vitest/globals", "jest"],
+    "types": [
+      "node",
+      "@testing-library/jest-dom",
+      "vitest/globals",
+      "jest",
+      "vite-plugin-pwa/client"
+    ],
     "target": "ESNext",
     "useDefineForClassFields": true,
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
